update: added changes done by niyati
These changes fix the bug that was being discussed on the #general channel. One issue/glitch that arises due to this is that when the rotatable object is not configured on 270 degrees and  if you tap on Cube S(x-5y8z3), the player will not move on this block but a block previous to that.