Make the obstacle to form a bridge to destination.
The obstacle falls to form a bridge to the destination after the collectable is collected by the player.

Conditionally activate the possible paths and obstacle blocks when needed.

Add and correct path elements of several blocks.

Various bugfixes and code cleanup.

Reduce the radius of the green collectable's collider.

Code cleanup and comments.